feat: move loadCategories request to the layout
